requirements.txt
pyspark>=3.5.0
boto3>=1.37.0
python-dotenv>=1.0.0
